+++ defaults/controllers/_default/all/_default/data.js
@@ -1,0 +1,18 @@
+// --------------------
+// Default controller
+// Default data action
+// --------------------
+
+// exports
+
+// action definition
+exports = module.exports = {
+    // functions
+
+    makeDisplayOptions: function(defaultFn) {
+        return defaultFn().bind(this)
+        .then(function() {
+            this.displayOptions.options.data = this.sequelize.Sequelize.getValues(this.data);
+        });
+    }
+};